+++ src/ballistica/shared/ballistica.cc
@@ -41,7 +41,7 @@ auto main(int argc, char** argv) -> int {
 namespace ballistica {
 
 // These are set automatically via script; don't modify them here.
-const int kEngineBuildNumber = 22510;
+const int kEngineBuildNumber = 22512;
 const char* kEngineVersion = "1.7.48";
 const int kEngineApiVersion = 9;